requirements.txt
# Document sync
httpx>=0.27.0
APScheduler>=3.10.0
python-dotenv>=1.0.0

# PDF loading (used by LangChain PyMuPDFLoader)
pymupdf>=1.24.0

# LangChain RAG stack
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0        # ChatOpenAI + OpenAIEmbeddings
langchain-anthropic>=0.2.0     # ChatAnthropic
langchain-chroma>=0.1.0        # Chroma vector store integration

# LLM / embedding providers
openai>=1.0.0
anthropic>=0.40.0

# Vector store
chromadb>=0.5.0
